+++ package.json
@@ -2,11 +2,11 @@
   "name": "chat-system",
   "version": "0.0.1",
   "description": "basic chat system",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "index": "node app/index.js",
-    "server": "nodemon server.js"
+    "server": "nodemon server.js",
+    "redis": "redis-cli"
   },
   "repository": {
     "type": "git",
@@ -21,6 +21,8 @@
   "dependencies": {
     "bootstrap": "5.2.0",
     "express": "4.18.1",
-    "nodemon": "2.0.19"
+    "nodemon": "2.0.19",
+    "redis": "4.2.0",
+    "websocket": "1.0.34"
   }
 }
