+++ server.js
@@ -2,11 +2,22 @@ const express = require('express');
 const app = express();
 const http = require('http');
 const server = http.createServer(app);
+var WebSocketServer = require('websocket').server;
 
 app.use(express.static("app"));
+var wsServer = new WebSocketServer({
+  httpServer: server
+});
+
+wsServer.on('request', function(request) {
+  var connection = request.accept(null, request.origin);
+  connection.on('message', (data) => {
+    connection.send(JSON.stringify(data.utf8Data));
+  });
+});
 
 app.get('/', (req, res) => {
-  res.sendFile(__dirname + '/app/index.html');
+  res.sendFile(__dirname + '/index.html');
 });
 
 server.listen(3000, () => {
